Bedre navngivning av beregningservicene
