+++ src/main/kotlin/no/nav/tilleggsstonader/sak/vedtak/dagligReise/KjørelisteSteg.kt
@@ -10,7 +10,7 @@ import no.nav.tilleggsstonader.sak.infrastruktur.exception.feil
 import no.nav.tilleggsstonader.sak.privatbil.avklartedager.AvklartKjørelisteService
 import no.nav.tilleggsstonader.sak.privatbil.avklartedager.finnesUkerMedAvvik
 import no.nav.tilleggsstonader.sak.vedtak.VedtakService
-import no.nav.tilleggsstonader.sak.vedtak.dagligReise.beregning.privatBil.PrivatBilBeregningsresultatService
+import no.nav.tilleggsstonader.sak.vedtak.dagligReise.beregning.privatBil.PrivatBilBeregningService
 import no.nav.tilleggsstonader.sak.vedtak.dagligReise.domain.BeregningsresultatPrivatBil
 import no.nav.tilleggsstonader.sak.vedtak.domain.InnvilgelseDagligReise
 import no.nav.tilleggsstonader.sak.vedtak.domain.InnvilgelseEllerOpphørDagligReise
@@ -18,7 +18,7 @@ import org.springframework.stereotype.Service
 
 @Service
 class KjørelisteSteg(
-    private val privatBilBeregningsresultatService: PrivatBilBeregningsresultatService,
+    private val privatBilBeregningService: PrivatBilBeregningService,
     private val vedtakService: VedtakService,
     private val arbeidsfordelingService: ArbeidsfordelingService,
     private val dagligReiseVedtakService: DagligReiseVedtakService,
@@ -47,7 +47,7 @@ class KjørelisteSteg(
                 ?: error("Finner ikke rammevedtak for behandling ${saksbehandling.id}")
 
         val beregningsresultatPrivatBil =
-            privatBilBeregningsresultatService.beregn(
+            privatBilBeregningService.beregn(
                 behandling = saksbehandling,
                 rammevedtak = eksisterendeRammevedtak,
                 brukersNavKontor = brukersNavKontor,
