+++ src/test/kotlin/no/nav/tilleggsstonader/sak/vedtak/dagligReise/beregning/privatBil/PrivatBilBeregningsresultatServiceTest.kt
@@ -44,7 +44,7 @@ class PrivatBilBeregningsresultatServiceTest {
     private val behandling = saksbehandling(id = behandlingId)
 
     val beregningService =
-        PrivatBilBeregningsresultatService(
+        PrivatBilBeregningService(
             avklartKjørelisteService = avklartKjørelisteService,
         )
